refactored auditlog backend-related code into its own separate file
