+++ frontend/src/components/AuditLogGrantVaultAccessDetails.vue
@@ -38,7 +38,8 @@
 <script setup lang="ts">
 import { onMounted, ref } from 'vue';
 import { useI18n } from 'vue-i18n';
-import { AuditLogEntityCache, AuthorityDto, GrantVaultAccessEventDto, VaultDto } from '../common/backend';
+import auditlog, { GrantVaultAccessEventDto } from '../common/auditlog';
+import { AuthorityDto, VaultDto } from '../common/backend';
 
 const { t } = useI18n({ useScope: 'global' });
 
@@ -46,14 +47,13 @@ const props = defineProps<{
   event: GrantVaultAccessEventDto
 }>();
 
-const entityCache = AuditLogEntityCache.getInstance();
 const resolvedUser = ref<AuthorityDto>();
 const resolvedVault = ref<VaultDto>();
 const resolvedAuthority = ref<AuthorityDto>();
 
 onMounted(async () => {
-  resolvedUser.value = await entityCache.getAuthority(props.event.userId);
-  resolvedVault.value = await entityCache.getVault(props.event.vaultId);
-  resolvedAuthority.value = await entityCache.getAuthority(props.event.authorityId);
+  resolvedUser.value = await auditlog.entityCache.getAuthority(props.event.userId);
+  resolvedVault.value = await auditlog.entityCache.getVault(props.event.vaultId);
+  resolvedAuthority.value = await auditlog.entityCache.getAuthority(props.event.authorityId);
 });
 </script>
