+++ frontend/src/components/AuditLogUpdateVaultMembershipDetails.vue
@@ -48,7 +48,8 @@
 <script setup lang="ts">
 import { onMounted, ref } from 'vue';
 import { useI18n } from 'vue-i18n';
-import { AuditLogEntityCache, AuthorityDto, UpdateVaultMembershipEventDto, VaultDto } from '../common/backend';
+import auditlog, { UpdateVaultMembershipEventDto } from '../common/auditlog';
+import { AuthorityDto, VaultDto } from '../common/backend';
 
 const { t } = useI18n({ useScope: 'global' });
 
@@ -56,14 +57,13 @@ const props = defineProps<{
   event: UpdateVaultMembershipEventDto
 }>();
 
-const entityCache = AuditLogEntityCache.getInstance();
 const resolvedUser = ref<AuthorityDto>();
 const resolvedVault = ref<VaultDto>();
 const resolvedAuthority = ref<AuthorityDto>();
 
 onMounted(async () => {
-  resolvedUser.value = await entityCache.getAuthority(props.event.userId);
-  resolvedVault.value = await entityCache.getVault(props.event.vaultId);
-  resolvedAuthority.value = await entityCache.getAuthority(props.event.authorityId);
+  resolvedUser.value = await auditlog.entityCache.getAuthority(props.event.userId);
+  resolvedVault.value = await auditlog.entityCache.getVault(props.event.vaultId);
+  resolvedAuthority.value = await auditlog.entityCache.getAuthority(props.event.authorityId);
 });
 </script>
